refactor(config): split monolith into logical files
Split internal/config/config.go into focused files:
- config_load.go: loading and defaults
- config_merge.go: profile and section merging
- config_validate.go: validation logic
- config_save.go: persistence

No functional changes; all exported APIs preserved.

Closes #171
